Declare almost all classes as Q_MOVABLE_TYPE
git-svn-id: svn://svn.code.sf.net/p/qtvlm/code/branches/qtVlm_dev@1404 2387b160-4621-4e8a-be36-40ea10d72da2

+++ src/Grib.cpp
@@ -368,8 +368,8 @@ void Grib::readAllGribRecords()
                                 }
                                 else {
                                         fprintf(stderr,
-                                                "GribReader: unknown record type: key=%Ld  idCenter==%d && idModel==%d && idGrid==%d\n",
-                                                rec->getKey(),
+                                                "GribReader: unknown record type: key=%d  idCenter==%d && idModel==%d && idGrid==%d\n",
+                                                (int)rec->getKey(),
                                                 rec->getIdCenter(), rec->getIdModel(), rec->getIdGrid()
                                                 );
                                 }
